factor: Move directory ent
