+++ go/ent/generate.go
