+++ go/ent/runtime/runtime.go
@@ -2,7 +2,7 @@
 
 package runtime
 
-// The schema-stitching logic is generated in GROWTHSPHERE/services/users/internal/infrastructure/database/ent/runtime.go
+// The schema-stitching logic is generated in GROWTHSPHERE/ent/runtime.go
 
 const (
 	Version = "v0.14.1"                                         // Version of ent codegen.
