+++ go/services/users/internal/domain/repository/userRepository.go
@@ -1,4 +1,5 @@
 package repository
 
 type IUserRepository interface {
+	
 }
